+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+
+ - Fixed a memory leak with vertex array objects.
+
 ## Version 0.1.1 (2015-03-13)
 
  - Added `ToColorAttachment` trait implementation for `Texture2dMultisample`.